Laravel 9.x Shift

+++ routes/backpack/backupmanager.php
@@ -1,5 +1,7 @@
 <?php
 
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | Backpack\BackupManager Routes
@@ -10,14 +12,10 @@
 |
 */
 
-Route::group([
-    'namespace'  => 'App\Http\Controllers\Admin',
-    'prefix'     => config('backpack.base.route_prefix', 'admin'),
-    'middleware' => ['web', config('backpack.base.middleware_key', 'admin')],
-], function () {
-    Route::get('backup', 'BackupController@index')->name('backup.index');
-    Route::put('backup/create', 'BackupController@create')->name('backup.store');
-    Route::put('backup/database/create', 'BackupController@createDatabase')->name('backup.database.store');
-    Route::get('backup/download/{file_name?}', 'BackupController@download')->name('backup.download');
-    Route::delete('backup/delete/{file_name?}', 'BackupController@delete')->where('file_name', '(.*)')->name('backup.destroy');
+Route::prefix(config('backpack.base.route_prefix', 'admin'))->middleware('web', config('backpack.base.middleware_key', 'admin'))->group(function () {
+    Route::get('backup', [App\Http\Controllers\Admin\BackupController::class, 'index'])->name('backup.index');
+    Route::put('backup/create', [App\Http\Controllers\Admin\BackupController::class, 'create'])->name('backup.store');
+    Route::put('backup/database/create', [App\Http\Controllers\Admin\BackupController::class, 'createDatabase'])->name('backup.database.store');
+    Route::get('backup/download/{file_name?}', [App\Http\Controllers\Admin\BackupController::class, 'download'])->name('backup.download');
+    Route::delete('backup/delete/{file_name?}', [App\Http\Controllers\Admin\BackupController::class, 'delete'])->where('file_name', '(.*)')->name('backup.destroy');
 });
